Merged in changes from RonDB 21.04, latency improvements, merged 8.0.31

+++ storage/ndb/clusterj/clusterj-core/src/main/java/com/mysql/clusterj/core/SessionImpl.java
@@ -152,6 +152,35 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
     private DTOCache dtoCache;
 
     boolean is_cached;
+
+    /**
+     * We need some support for O(1) handling of LRU list of Session objects
+     * also in the presence of multiple database Session objects.
+     */
+    SessionImpl next_lru_list;
+    SessionImpl prev_lru_list;
+
+    SessionImpl getNextLruList() {
+        return next_lru_list;
+    }
+    SessionImpl getPrevLruList() {
+        return prev_lru_list;
+    }
+    void setNextLruList(SessionImpl session) {
+        next_lru_list = session;
+    }
+    void setPrevLruList(SessionImpl session) {
+        prev_lru_list = session;
+    }
+
+    String getDatabaseName() {
+        return db.getName();
+    }
+
+    boolean isDefaultDatabase() {
+        return db.isDefaultDatabase();
+    }
+
     /** Create a SessionImpl with factory, properties, Db, and dictionary
      */
     SessionImpl(SessionFactoryImpl factory, Map properties, Db db, Dictionary dictionary,
@@ -641,7 +670,7 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
         int count = 0;
         try {
             op = clusterTransaction.getTableScanOperationLockModeExclusiveScanFlagKeyInfo(storeTable);
-            count = deletePersistentAll(op, true);
+            count = deletePersistentAll(op, true, 0, Long.MAX_VALUE);
         } catch (ClusterJException ex) {
             failAutoTransaction();
             // TODO add table name to the error message
@@ -657,23 +686,40 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
      * @param op the scan operation
      * @return the number of instances deleted
      */
-    public int deletePersistentAll(ScanOperation op, boolean abort) {
+    public int deletePersistentAll(ScanOperation op,
+                                   boolean abort,
+                                   long skip,
+                                   long limit) {
         int cacheCount = 0;
         int count = 0;
+        int delete_count = 0;
         boolean done = false;
         boolean fetch = true;
         // cannot use early autocommit optimization here
         clusterTransaction.setAutocommit(false);
+        // check that limit is not zero, if so we're done
+        assert(limit > 0);
         // execute the operation
         clusterTransaction.executeNoCommit(true, true);
         while (!done ) {
             int result = op.nextResult(fetch);
             switch (result) {
                 case RESULT_READY:
-                    op.deleteCurrentTuple();
+                    if (skip <= 0 || count >= skip) {
+                        op.deleteCurrentTuple();
+                        ++cacheCount;
+                        ++delete_count;
+                        fetch = false;
+                        if (delete_count == limit) {
+                            done = true;
+                            if (cacheCount != 0) {
+                                clusterTransaction.executeNoCommit(abort, true);
+                            }
+                            cacheCount = 0;
+                            op.close();
+                        }
+                    }
                     ++count;
-                    ++cacheCount;
-                    fetch = false;
                     break;
                 case SCAN_FINISHED:
                     done = true;
@@ -683,7 +729,9 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
                     op.close();
                     break;
                 case CACHE_EMPTY:
-                    clusterTransaction.executeNoCommit(abort, true);
+                    if (cacheCount != 0) {
+                        clusterTransaction.executeNoCommit(abort, true);
+                    }
                     cacheCount = 0;
                     fetch = true;
                     break;
@@ -692,7 +740,7 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
                             local.message("ERR_Next_Result_Illegal", result));
             }
         }
-        return count;
+        return delete_count;
     }
 
     /** Select a single row from the database. Only the fields requested
@@ -876,7 +924,11 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
         if (dtoCache != null && dropCache) {
             dtoCache.drop();
         }
-        factory.storeCachedSession(this);
+        if (db.isDefaultDatabase()) {
+            factory.storeCachedSession(this);
+        } else {
+            factory.storeCachedSession(this, db.getName());
+        }
     }
 
     public void setCached(boolean is_cached_val) {
@@ -1660,7 +1712,7 @@ public class SessionImpl implements SessionSPI, CacheManager, StoreManager {
         assertNotClosed();
         //drop all caches as they are invalid now due to change in schema ID
         dropInstanceCache();
-        return factory.unloadSchema(cls, dictionary);
+        return factory.unloadSchema(cls, dictionary, db.getName(), db.isDefaultDatabase());
     }
 
     /** Release resources associated with an instance. The instance must be a domain object obtained via
